+++ src/llm/async_inference_engine.cpp
@@ -689,7 +689,7 @@ json AsyncInferenceEngine::getWorkerStats() const {
             std::sort(sorted.begin(), sorted.end());
             size_t p99_idx = static_cast<size_t>(sorted.size() * 0.99);
             stats["p99_latency_ms"] =
-                sorted[std::min(p99_idx, static_cast<int>(sorted.size()) - 1)];
+                sorted[std::min(p99_idx, sorted.size() - 1)];
         }
     }
